\name{cholesterol}
\alias{cholesterol}
\docType{data}
\title{cholesterol of a imagimary class}
\description{

}
\usage{data(cholesterol)}
\format{
  A data frame with 4 observations bservations on the following 5 variables.
  This data is a summary from a simple random sample of 100 individuals.
  \describe{
    \item{agegroup}{Age group}
    \item{samplesize}{Sample size in age group}
    \item{chol.mean}{mean of the cholesterol rate over the respondents}
    \item{chol.sd}{standard deviation of the cholesterol rate over the respondents}  
    \item{nresp}{Number of respondents in age group}
    \item{samplesize}{Sample size in age group}}
}

\source{Statistical Analysis with Missing data, Little and Rubin, 2nd Ed. p.~56 }
\examples{
data(cholesterol)
}
\keyword{datasets}
